ver-claude/requirements.txt
click>=8.0.0
requests>=2.25.0
feedparser>=6.0.0
python-dateutil>=2.8.0
